[None][perf] idea_0067_v2_perf_v1: store DSV4 lm_head weight in fp32 to hoist the per-step bf16->fp32 recast (idea idea_0067)
DeepseekV4ForCausalLM.forward recomputed self.head.weight.float() every decode
step -- a constant [vocab/8, hidden] = 66M-element bf16->fp32 cast that cudagraph
replays once per token (10 kernels in the proxy decode window, ~178us each).
head.weight is a constant parameter, so the cast reproduces the same fp32 tensor
each step. This constructs self.head as nn.Linear(..., dtype=torch.float32) and
drops the per-step .float() in the lm_head linear. _load_checkpoint copy-casts
the bf16 checkpoint into the fp32 param (no assign=True), so the stored values
equal bf16(W).float() and the logits GEMM is bit-identical in the real flow.
shard_weight_tensor / _load_hook are dtype-agnostic (slice only), so idea_0059's
gather-shard of the head survives unchanged (fp32 shards, fp32 all_gather).

DeepSeek-V4-Flash proxy(10L) B200 EP=8 decode, copy_cast op-type share:
  raw      pre 29.05% -> post 28.28% / 30.04% (comm-noise-masked: NCCL comm
           swung 16.68%->15.51%->10.10% across runs, a 6.6pp denominator swing
           that inverts the raw share sign -- unusable near the floor)
  ex-comm  pre 34.87% -> post 33.47% / 33.42% = -1.40/-1.45pp (2 edited runs,
           reproducible, clears 1.0pp SHARE_FLOOR)
  instances 13680 -> 13670 (-10, deterministic: the 10 per-step head casts)
  copy_cast ABS 35483us -> 33224/33129us (-2259/-2353us); the deterministic
  head-cast component is 1780us = 1.46% of proxy decode GPU time.
Full-model tpot confirmation deferred to Tier-2 batched validation (head cast is
a fixed ~178us/step, ~0.59% of the 30.078ms tpot reference).

Accuracy (test_deepseek_v4_lm_head_fp32, 2/2 passed): fp32-stored head is
torch.equal to the per-step bf16->fp32 recast (bf16->fp32 is lossless), i.e.
bit-identical in the checkpoint-loaded flow. The FAST_MODE proxy outputs diverge
only because random-init allocates the fp32 head directly as fp32-normal rather
than bf16-rounded-normal -- a random-init artifact, not a real-model divergence.

+++ tensorrt_llm/_torch/auto_deploy/models/custom/modeling_deepseek_v4.py
@@ -1668,7 +1668,14 @@ class DeepseekV4ForCausalLM(DeepseekV4PreTrainedModel, GenerationMixin):
             [DeepseekV4Block(config, layer_idx) for layer_idx in range(config.num_hidden_layers)]
         )
         self.norm = DeepseekV4RMSNorm(config.hidden_size, config.rms_norm_eps)
-        self.head = nn.Linear(config.hidden_size, config.vocab_size, bias=False)
+        # Store the LM-head weight in fp32 so the per-step bf16->fp32 recast in
+        # ``forward`` (``self.head.weight.float()``) is hoisted to a one-time
+        # load-time cast. ``_load_checkpoint`` copy-casts the bf16 checkpoint
+        # tensor into this fp32 param (no ``assign=True``), so the stored values
+        # equal ``bf16(W).float()`` and the logits GEMM is bit-identical.
+        self.head = nn.Linear(
+            config.hidden_size, config.vocab_size, bias=False, dtype=torch.float32
+        )
         self.rotary_emb = DeepseekV4RotaryEmbedding(config)
         self.hc_mult = config.hc_mult
         hc_dim = config.hc_mult * config.hidden_size
@@ -1723,8 +1730,10 @@ class DeepseekV4ForCausalLM(DeepseekV4PreTrainedModel, GenerationMixin):
         # every rank re-casts and matmuls the full [vocab, hidden] weight each
         # step; sharded, both the bf16->fp32 weight cast and the fp32 logits GEMM
         # shrink by tp_size. Mathematically exact (output columns are independent).
+        # ``self.head.weight`` is already fp32 (cast once at load), so no per-step
+        # bf16->fp32 weight recast is emitted into the decode graph here.
         logits = _linear(
-            hidden_states.float(), self.head.weight.float(), None, layer_type="lm_head"
+            hidden_states.float(), self.head.weight, None, layer_type="lm_head"
         ).float()
         return DeepseekV4CausalLMOutput(logits=logits)
